Move dump_operation to utils module
