+++ pip/qsharp/utils/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+
+from ._utils import dump_operation
+
+__all__ = [
+    "dump_operation",
+]
