+++ pip/tests/test_qsharp.py
@@ -2,6 +2,7 @@
 # Licensed under the MIT License.
 
 import qsharp
+import qsharp.utils
 from contextlib import redirect_stdout
 import io
 
@@ -80,14 +81,14 @@ def test_dump_machine() -> None:
 
 def test_dump_operation() -> None:
     qsharp.init(target_profile=qsharp.TargetProfile.Unrestricted)
-    res = qsharp.dump_operation("qs => ()", 1)
+    res = qsharp.utils.dump_operation("qs => ()", 1)
     assert res == [[complex(1.0, 0.0), complex(0.0, 0.0)],
                    [complex(0.0, 0.0), complex(1.0, 0.0)]]
-    res = qsharp.dump_operation("qs => H(qs[0])", 1)
+    res = qsharp.utils.dump_operation("qs => H(qs[0])", 1)
     assert res == [[complex(0.707107, 0.0), complex(0.707107, 0.0)],
                    [complex(0.707107, 0.0), complex(-0.707107, 0.0)]]
     qsharp.eval("operation ApplySWAP(qs : Qubit[]) : Unit { SWAP(qs[0], qs[1]); }")
-    res = qsharp.dump_operation("ApplySWAP", 2)
+    res = qsharp.utils.dump_operation("ApplySWAP", 2)
     assert res == [[complex(1.0, 0.0), complex(0.0, 0.0), complex(0.0, 0.0), complex(0.0, 0.0)],
                    [complex(0.0, 0.0), complex(0.0, 0.0), complex(1.0, 0.0), complex(0.0, 0.0)],
                    [complex(0.0, 0.0), complex(1.0, 0.0), complex(0.0, 0.0), complex(0.0, 0.0)],
